camera: infuse4g: Fix "zoomed in" FFC
This sets the PREVIEW_ONOFF control for the FFC on Infuse,
which prevents it from getting zoomed in.

TODO: Fix capture of 1280x960 stills.  640x480 is all
that is supported by this code.  Attempting to implement
1280x960 causes the camera app to crash.

It also alters SecCamera.h to use alternate #defines
for each camera type the way the code was intended
instead of just #ifdefing the focal length.

Still much work to be done on Infuse FFC...  There are
a number of controls this HAL attempts to use that
are unsupported by SR130PC10.  Consider this a temporary
workaround.

Change-Id: Ic978e76539e4d4c681e49cd8b10b517d98b4d2e3