Change order of flash "current state" change.
In the current setup, if the flash mode has changed,
it will set the "current" state to the "next" state,
even if it turns out that it cannot actually change
the state.

This makes the "current" flash state only update
when the flash state is actually changed with
a v4l2 call.

Change-Id: I6d1e2f3eccf4054fa25d6c991aae7840184eec6a